requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
pyparsing==2.4.6
PySocks==1.7.1
python-telegram-bot==13.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
schedule==0.6.0
six==1.14.0
tornado==6.1
tweepy==3.9.0
tzlocal==2.1
urllib3==1.25.8
webencodings==0.5.1
